+++ CHANGELOG.md
@@ -1,2 +1,8 @@
 # Change Log
 
+## 0.0.1
+Initial Version
+* ADD. command `enumerator.fill-enum-ex` for Insert by format `[0|x]<start> [<step>]`
+* ADD. command `enumerator.fill-enum-dec` for Insert dec values
+* ADD. command `enumerator.fill-enum-dec-padded` for Insert zero-padded dec values
+* ADD. command `enumerator.fill-enum-hex` for Insert hex values